+++ sdk/python/tests/integration/e2e/test_go_feature_server.py
@@ -35,97 +35,6 @@ from tests.utils.feature_utils import generate_expected_logs, get_latest_rows
 from tests.utils.http_utils import check_port_open, free_port
 
 
-def _server_port(environment, server_type: str):
-    if not environment.test_repo_config.go_feature_serving:
-        pytest.skip("Only for Go path")
-
-    fs = environment.feature_store
-
-    embedded = EmbeddedOnlineFeatureServer(
-        repo_path=str(fs.repo_path.absolute()),
-        repo_config=fs.config,
-        feature_store=fs,
-    )
-    port = free_port()
-    if server_type == "grpc":
-        target = embedded.start_grpc_server
-    elif server_type == "http":
-        target = embedded.start_http_server
-    else:
-        raise ValueError("Server Type must be either 'http' or 'grpc'")
-
-    t = threading.Thread(
-        target=target,
-        args=("127.0.0.1", port),
-        kwargs=dict(
-            enable_logging=True,
-            logging_options=FeatureLoggingConfig(
-                enabled=True,
-                queue_capacity=100,
-                write_to_disk_interval_secs=1,
-                flush_interval_secs=1,
-                emit_timeout_micro_secs=10000,
-            ),
-        ),
-    )
-    t.start()
-
-    wait_retry_backoff(
-        lambda: (None, check_port_open("127.0.0.1", port)), timeout_secs=15
-    )
-
-    yield port
-    if server_type == "grpc":
-        embedded.stop_grpc_server()
-    else:
-        embedded.stop_http_server()
-
-    # wait for graceful stop
-    time.sleep(5)
-
-
-# Go test fixtures
-
-
-@pytest.fixture
-def initialized_registry(environment, universal_data_sources):
-    fs = environment.feature_store
-
-    _, _, data_sources = universal_data_sources
-    feature_views = construct_universal_feature_views(data_sources)
-
-    feature_service = FeatureService(
-        name="driver_features",
-        features=[feature_views.driver],
-        logging_config=LoggingConfig(
-            destination=environment.data_source_creator.create_logged_features_destination(),
-            sample_rate=1.0,
-        ),
-    )
-    feast_objects: List[FeastObject] = [feature_service]
-    feast_objects.extend(feature_views.values())
-    feast_objects.extend([driver(), customer(), location()])
-
-    fs.apply(feast_objects)
-    fs.materialize(environment.start_date, environment.end_date)
-
-
-@pytest.fixture
-def grpc_server_port(environment, initialized_registry):
-    yield from _server_port(environment, "grpc")
-
-
-@pytest.fixture
-def http_server_port(environment, initialized_registry):
-    yield from _server_port(environment, "http")
-
-
-@pytest.fixture
-def grpc_client(grpc_server_port):
-    ch = grpc.insecure_channel(f"localhost:{grpc_server_port}")
-    yield ServingServiceStub(ch)
-
-
 @pytest.mark.integration
 @pytest.mark.goserver
 def test_go_grpc_server(grpc_client):
@@ -255,3 +164,99 @@ def test_feature_logging(
     persisted_logs = persisted_logs.sort_values(by="driver_id").reset_index(drop=True)
     persisted_logs = persisted_logs[expected_logs.columns]
     pd.testing.assert_frame_equal(expected_logs, persisted_logs, check_dtype=False)
+
+
+"""
+Start go feature server either on http or grpc based on the repo configuration for testing.
+"""
+
+
+def _server_port(environment, server_type: str):
+    if not environment.test_repo_config.go_feature_serving:
+        pytest.skip("Only for Go path")
+
+    fs = environment.feature_store
+
+    embedded = EmbeddedOnlineFeatureServer(
+        repo_path=str(fs.repo_path.absolute()),
+        repo_config=fs.config,
+        feature_store=fs,
+    )
+    port = free_port()
+    if server_type == "grpc":
+        target = embedded.start_grpc_server
+    elif server_type == "http":
+        target = embedded.start_http_server
+    else:
+        raise ValueError("Server Type must be either 'http' or 'grpc'")
+
+    t = threading.Thread(
+        target=target,
+        args=("127.0.0.1", port),
+        kwargs=dict(
+            enable_logging=True,
+            logging_options=FeatureLoggingConfig(
+                enabled=True,
+                queue_capacity=100,
+                write_to_disk_interval_secs=1,
+                flush_interval_secs=1,
+                emit_timeout_micro_secs=10000,
+            ),
+        ),
+    )
+    t.start()
+
+    wait_retry_backoff(
+        lambda: (None, check_port_open("127.0.0.1", port)), timeout_secs=15
+    )
+
+    yield port
+    if server_type == "grpc":
+        embedded.stop_grpc_server()
+    else:
+        embedded.stop_http_server()
+
+    # wait for graceful stop
+    time.sleep(5)
+
+
+# Go test fixtures
+
+
+@pytest.fixture
+def initialized_registry(environment, universal_data_sources):
+    fs = environment.feature_store
+
+    _, _, data_sources = universal_data_sources
+    feature_views = construct_universal_feature_views(data_sources)
+
+    feature_service = FeatureService(
+        name="driver_features",
+        features=[feature_views.driver],
+        logging_config=LoggingConfig(
+            destination=environment.data_source_creator.create_logged_features_destination(),
+            sample_rate=1.0,
+        ),
+    )
+    feast_objects: List[FeastObject] = [feature_service]
+    feast_objects.extend(feature_views.values())
+    feast_objects.extend([driver(), customer(), location()])
+
+    fs.apply(feast_objects)
+    fs.materialize(environment.start_date, environment.end_date)
+
+
+@pytest.fixture
+def grpc_server_port(environment, initialized_registry):
+    yield from _server_port(environment, "grpc")
+
+
+@pytest.fixture
+def http_server_port(environment, initialized_registry):
+    yield from _server_port(environment, "http")
+
+
+@pytest.fixture
+def grpc_client(grpc_server_port):
+    ch = grpc.insecure_channel(f"localhost:{grpc_server_port}")
+    yield ServingServiceStub(ch)
